Refactor/minor updates

+++ app/Http/Controllers/LikeController.php
@@ -29,14 +29,18 @@ class LikeController extends Controller
 		return response()->json(['message' => 'User Liked Post', 'like' => LikeResource::make($like)], 201);
 	}
 
-	public function destroy(): JsonResponse
+	public function destroy(Quote $quote): JsonResponse
 	{
-		$like = Like::where('user_id', auth()->id())->first();
+		$like = $quote->likes()
+			->where('user_id', auth()->id())
+			->first();
 
-		event(new ReactPost(LikeResource::make($like)));
+		if ($like) {
+			event(new ReactPost(LikeResource::make($like)));
 
-		$like->delete();
-
-		return response()->json(['message' => 'User Unliked Post'], 204);
+			$like->delete();
+			return response()->json(['message' => 'User Unliked Post'], 204);
+		}
+		return response()->json(['message' => 'Like not found for the quote'], 400);
 	}
 }
